Overhaul (#349)
* Mini GPU overhaul

* Alch value

* Stoned Loot Tracker (persistent loot tracker)

* Checkstyle

* Added back max values to GPU plugin

* Persistent loot tracker config options

* Zulrah plugin type

* Add inventory highlight plugin

When dragging an item in the inventory, a highlight will show up to help
the player place the item in the correct slot, instead of missing when
they drag the item to the "gutter"

* Fix inventory highlight plugin

* Inventory highlight plugin type

* Inventory highlight config options

* Fix ess pouch swapping

* Runecraftig config options

* Add plugin type to persistient loot tracker

* Fix panel repaint

* Disable this for now

* Revert "Fix ess pouch swapping" -> PR #351

This reverts commit c18f65069f5df3a1cafb983af5b3e162817a22fe.

* I blame intellij for this

* Fix test

+++ runelite-client/src/main/java/net/runelite/client/plugins/grandexchange/GrandExchangeItemPanel.java
@@ -52,7 +52,7 @@ class GrandExchangeItemPanel extends JPanel
 {
 	private static final Dimension ICON_SIZE = new Dimension(32, 32);
 
-	GrandExchangeItemPanel(AsyncBufferedImage icon, String name, int itemID, int gePrice, Double
+	GrandExchangeItemPanel(AsyncBufferedImage icon, String name, int itemID, int gePrice, int
 		haPrice, int geItemLimit)
 	{
 		BorderLayout layout = new BorderLayout();
@@ -139,7 +139,7 @@ class GrandExchangeItemPanel extends JPanel
 
 		// Alch price
 		JLabel haPriceLabel = new JLabel();
-		haPriceLabel.setText(StackFormatter.formatNumber(haPrice.intValue()) + " alch");
+		haPriceLabel.setText(StackFormatter.formatNumber(haPrice) + " alch");
 		haPriceLabel.setForeground(ColorScheme.GRAND_EXCHANGE_ALCH);
 		alchAndLimitPanel.add(haPriceLabel, BorderLayout.WEST);
